+++ Surface_mesh_segmentation/include/CGAL/internal/Surface_mesh_segmentation/Expectation_maximization.h
@@ -214,7 +214,7 @@ private:
         double distance = std::abs(centers[i].mean - *it);
         if(distance < min_distance) {
           min_distance = distance;
-          closest_center = i;
+          closest_center = static_cast<int>(i);
         }
       }
       member_count[closest_center]++;
